Updated some pages by adding prevent direct url access

+++ shopping/delete-all-item.php
@@ -2,9 +2,16 @@
 <?php require "../config/config.php"; ?>
 
 <?php
+if ($_SERVER['REQUEST_METHOD'] == 'GET' && realpath(__FILE__) == realpath($_SERVER['SCRIPT_FILENAME'])) {
+    header('HTTP/1.0 403 Forbidden', TRUE, 403);
+    die(header('location: http://localhost:3000/bookstore'));
+}
+?>
 
-if(isset($_POST['delete'])){
-  
+<?php
+
+if (isset($_POST['delete'])) {
+
     $delete = $conn->prepare("DELETE FROM cart WHERE user_id='$_SESSION[user_id]'");
     $delete->execute();
 }
